+++ src/main/java/com/github/ucchyocean/lc3/command/ListCommand.java
@@ -5,20 +5,20 @@
  */
 package com.github.ucchyocean.lc3.command;
 
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.Comparator;
-
 import com.github.ucchyocean.lc3.Messages;
 import com.github.ucchyocean.lc3.channel.Channel;
 import com.github.ucchyocean.lc3.member.ChannelMember;
 import com.github.ucchyocean.lc3.util.ChatColor;
-
 import net.md_5.bungee.api.chat.BaseComponent;
 import net.md_5.bungee.api.chat.TextComponent;
 
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.Comparator;
+
 /**
  * listコマンドの実行クラス
+ *
  * @author ucchy
  */
 public class ListCommand extends LunaChatSubCommand {
@@ -30,6 +30,7 @@ public class ListCommand extends LunaChatSubCommand {
 
     /**
      * コマンドを取得します。
+     *
      * @return コマンド
      * @see com.github.ucchyocean.lc3.command.LunaChatSubCommand#getCommandName()
      */
@@ -40,6 +41,7 @@ public class ListCommand extends LunaChatSubCommand {
 
     /**
      * パーミッションノードを取得します。
+     *
      * @return パーミッションノード
      * @see com.github.ucchyocean.lc3.command.LunaChatSubCommand#getPermissionNode()
      */
@@ -50,6 +52,7 @@ public class ListCommand extends LunaChatSubCommand {
 
     /**
      * コマンドの種別を取得します。
+     *
      * @return コマンド種別
      * @see com.github.ucchyocean.lc3.command.LunaChatSubCommand#getCommandType()
      */
@@ -60,8 +63,9 @@ public class ListCommand extends LunaChatSubCommand {
 
     /**
      * 使用方法に関するメッセージをsenderに送信します。
+     *
      * @param sender コマンド実行者
-     * @param label 実行ラベル
+     * @param label  実行ラベル
      * @see com.github.ucchyocean.lc3.command.LunaChatSubCommand#sendUsageMessage()
      */
     @Override
@@ -72,9 +76,10 @@ public class ListCommand extends LunaChatSubCommand {
 
     /**
      * コマンドを実行します。
+     *
      * @param sender コマンド実行者
-     * @param label 実行ラベル
-     * @param args 実行時の引数
+     * @param label  実行ラベル
+     * @param args   実行時の引数
      * @return コマンドが実行されたかどうか
      * @see com.github.ucchyocean.lc3.command.LunaChatSubCommand#runCommand(java.lang.String[])
      */
@@ -83,7 +88,7 @@ public class ListCommand extends LunaChatSubCommand {
             ChannelMember sender, String label, String[] args) {
 
         int page = 1;
-        if ( args.length >= 2 && args[1].matches("[0-9]+") ) {
+        if (args.length >= 2 && args[1].matches("[0-9]+")) {
             page = Integer.parseInt(args[1]);
         }
 
@@ -96,21 +101,22 @@ public class ListCommand extends LunaChatSubCommand {
 
     /**
      * リスト表示用のリストを返す
+     *
      * @param player プレイヤー、指定しない場合はnullにする
-     * @param page 表示するページ、0を指定した場合は全表示
+     * @param page   表示するページ、0を指定した場合は全表示
      * @return リスト
      */
     private ArrayList<BaseComponent[]> getList(ChannelMember player, int page) {
 
         ArrayList<BaseComponent[]> list = getPlayerList(player);
         int size = list.size();
-        int maxPage = (int)(size / PAGE_SIZE) + 1;
+        int maxPage = (size / PAGE_SIZE) + 1;
 
-        if ( page < 0 ) page = 0;
-        if ( page > maxPage ) page = maxPage;
+        if (page < 0) page = 0;
+        if (page > maxPage) page = maxPage;
 
         ArrayList<BaseComponent[]> items = new ArrayList<>();
-        if ( page == 0 ) { // 全表示
+        if (page == 0) { // 全表示
             items.add(TextComponent.fromLegacyText(Messages.listFirstLine()));
             items.addAll(list);
             items.add(TextComponent.fromLegacyText(Messages.listEndLine()));
@@ -126,6 +132,7 @@ public class ListCommand extends LunaChatSubCommand {
 
     /**
      * 指定されたプレイヤーに対するチャンネルリストを返す
+     *
      * @param player プレイヤー
      * @return チャンネルリスト
      */
@@ -133,10 +140,10 @@ public class ListCommand extends LunaChatSubCommand {
 
         String dchannel = "";
         String playerName = "";
-        if ( player != null ) {
+        if (player != null) {
             playerName = player.getName();
             Channel def = api.getDefaultChannel(playerName);
-            if ( def == null ) {
+            if (def == null) {
                 dchannel = "";
             } else {
                 dchannel = def.getName();
@@ -147,43 +154,43 @@ public class ListCommand extends LunaChatSubCommand {
         ArrayList<Channel> channels = new ArrayList<>(api.getChannels());
         Collections.sort(channels, new Comparator<Channel>() {
             public int compare(Channel c1, Channel c2) {
-                if ( c1.getOnlineNum() == c2.getOnlineNum() ) return c1.getName().compareTo(c2.getName());
+                if (c1.getOnlineNum() == c2.getOnlineNum()) return c1.getName().compareTo(c2.getName());
                 return c2.getOnlineNum() - c1.getOnlineNum();
             }
         });
 
         // 指定されたプレイヤー名に合うように、フィルタ＆表示用整形する。
         ArrayList<BaseComponent[]> items = new ArrayList<>();
-        for ( Channel channel : channels ) {
+        for (Channel channel : channels) {
 
             // BANされているチャンネルは表示しない
-            if ( channel.getBanned().contains(player) ) {
+            if (channel.getBanned().contains(player)) {
                 continue;
             }
 
             // 個人チャットはリストに表示しない
-            if ( channel.isPersonalChat() ) {
+            if (channel.isPersonalChat()) {
                 continue;
             }
 
             // パスワードが設定されているチャンネルはリストに表示しない
-            if( !player.hasPermission("lunachat-admin.list-all-channels") && !channel.getPassword().equals("") ) {
+            if (!player.hasPermission("lunachat-admin.list-all-channels") && !channel.getPassword().equals("")) {
                 continue;
             }
 
             // デフォルト発言先なら赤に、非表示中なら暗青にする。
             String disp = ChatColor.WHITE + channel.getName();
-            if ( channel.getName().equalsIgnoreCase(dchannel) ) {
+            if (channel.getName().equalsIgnoreCase(dchannel)) {
                 disp = ChatColor.RED + channel.getName();
-            } else if ( channel.getHided().contains(player) ) {
+            } else if (channel.getHided().contains(player)) {
                 disp = ChatColor.DARK_AQUA + channel.getName();
             }
 
-            if ( !channel.getMembers().contains(player) &&
-                    !channel.isGlobalChannel() ) {
+            if (!channel.getMembers().contains(player) &&
+                    !channel.isGlobalChannel()) {
 
                 // 未参加で visible=false のチャンネルは表示しない
-                if ( !channel.isVisible() ) {
+                if (!channel.isVisible()) {
                     continue;
                 }
                 disp = ChatColor.GRAY + channel.getName();
